C:raid: update hpsa logic to install ssacli tools on > buster
We currently have on other instance of using the ssacli tools which is
hard coded based on the raid fact looking for a specific pci id[1].

This is a bit of a hack and i dont think it gives the correct picture the
raid driver is still (i think hpsa) its just that we need newer tools.

It also seems that on bullseye we also need to use the ssacli tools.
this CR updates the code so that the entry point is raid::hpsa which
then installs the correct set of tools.  currently this is just based on
the os version.  however we should also handle the case we currently
have.  For now i have just added some notes and some backwards
compatibility classes

i also wonder if we could just move everything to just use ssacli?


[1]https://gerrit.wikimedia.org/r/c/operations/puppet/+/503332/6/modules/raid/lib/facter/raid.rb

Bug: T306354
Hosts: C:raid
Change-Id: I10428ce6d0e3d9842a590acc86d1ff385f1f822b