P:cache::varnish::frontend: mask the varnishncsa service
We don't use the varnishncsa service and we prevent it from starting by
ensuring that it is stopped. However, this doesn't seem to be enough as
there are cases (especially after a reboot), where the Puppet run fails
because the service has been started. An example of such a failure is
[0], where the reimage cookbook stalled because it couldn't find a
successful Puppet run.

The solution was to reset-failed varnishncsa and then manually run
Puppet, but that isn't desirable so let's mask the service so that it
cannot be started by accident.

While we are at it, also mask varnishlog since it's not been used.

[0] - https://puppetboard.wikimedia.org/node/cp5002.eqsin.wmnet

Hosts: cp5002.eqsin.wmnet
Change-Id: I3fe2c3a8d960d9f0a23e9a2f0b26ba36e1ae46e4