durum: add monitoring::service for the check service
Add a monitoring::service for durum that primarily checks if nginx is
active and listening on the anycasted IP. Ideally, this check should be
run against the host IP and not the anycast IP, but for durum we don't
listen on the host IP on nginx so that doesn't work for us. (The
expectation with this commit being that a durum host when it tries to
reach the anycast IP actually reaches itself.)

The anycast IP monitoring has already been added in
profile::bird::anycast_monitoring.

Change-Id: I8e8b75390066076bc640d7fa773c6bc7f40b32c9
Hosts: O:durum