puppetmaster::geoip: remove legacy product IDs even for fallback option
We have already removed the legacy product IDs when we "fetch_private"
on the puppetmaster.

But there is a fallback option when fetch_private is set to false.
This is the case in deployment-prep in cloud VPS. (beta).

Since we are getting reminders the legacy IDs need to be removed
let's also do that for the fallback case..

Bug: T302864
Change-Id: I7453fea137643b605c937f38dcaacec49bf80156