kubernetes: allow deploy-ml-service users to check pods on ml-serve
The ML-team members are not in the deployment group, since they
don't need those permissions. It seems that to access the kube_env
credentials to inspect ml-serve pods one needs to be inside
deployment user group, but this seems to be unnecessary for
the ml-serve use case (maybe in the future but probably not now).

The alternative is to add ML folks to the deployment user group,
what people think it is best.

Change-Id: Id8e16ff6cd844d3385e43ae91131cba619cd372f