Rewrite logster::job to use systemd timers.
This patch does change the meaning of the variable "weekday", going from being a number
between 0 and 6, to a string. Weekday is currently never used (only minutes are).
The change will only affect patches and roles that have yet to be written.

Redirect of stdout and stderr has been removed to allow Journald to pick up output,
which can be useful for debugging.

Bug: T273673
Change-Id: I8a1fa5465b93ce32e2f477b54a6ef23e3f79c13d