zuul: disable core.logAllRefUpdates at clone time
To prevent the git reflog to be generated on the Zuul merger,
fa07ec931b9 had set `core.logAllRefUpdates`to `True` in the git user
configuration.

However, when git clones a non-bare repository, it crafts a
configuration local to the repository which sets `core.logAllRefUpdates`
to `True` which takes precedence over the user configuration:

  $ sudo -H -u zuul git config --list
  core.logallrefupdates=false
  core.logallrefupdates=true

  $ git config --list --show-origin|grep -i logallref
  file:/var/lib/zuul/.gitconfig	core.logallrefupdates=false
  file:.git/config	core.logallrefupdates=true

`git-init` (which is invoked by `git-clone`) can copy a default config
from a directory and has the logic to recognizes when
`core.logAllRefUpdates` has been overriden in such a way:

builtin/init-db.c
    if (is_bare_repository())
            git_config_set("core.bare", "true");
    else {
            git_config_set("core.bare", "false");
 ---->      /* allow template config file to override the default */
            if (log_all_ref_updates == LOG_REFS_UNSET)
                    git_config_set("core.logallrefupdates", "true");

This change create a `git-init` config template under the user directory
and configure `git-init` to use it.

Bug: T307620
Hosts: contint2001.wikimedia.org,contint1001.wikimedia.org
Change-Id: I5b4869e7a606f392585056358e7be7a31721da87