Netbox: update config file for 3.1
Netbox introduced a feature named "dynamic configuration", where
configuration knobs are stored in the DB instead of config files.
This patch removes the variables that are not used anymore.

This will needs to be merged once prod is running 3.1 and not before.
Keeping the variables on 3.1 or above should be a noop.

Those 3 variables are different from their default setting:
MAX_PAGE_SIZE = None (default: 1000)
PAGINATE_COUNT = 1000 (default: 50)
CHANGELOG_RETENTION = 730 (default: 90)

The last one might cause an issue where the SCAP provisioning script
cleans entries older than 90 days before we have time to access the UI
and change the default.
Edit: at least on 3.1 the hardcoded values override the DB ones,
so it's not an issue. To be tested on 3.2.

Note that 3.2 doesn't have any relevant change to its config file.

Bug: T296452
Change-Id: I4f969b4a7aa335190e71816d15ea89f4c0d30e20