swift: drain ms-be1059, skip cluster-OK checks
ms-be1059 is currently bricked, awaiting assistance from the
vendor. In the mean time, we can't do any other upgrades of this
cluster safely, and write latency has gone up.

So drain it; do this somewhat gradually rather than just
immediately. But to do that we have to skip the usual cluster health
checks (which the ring manager does to check that a previous change
has completed before making further updates). The effect of this will
be to drain the host over the next 50 hours or so.

Bug: T307667
Change-Id: Ib10196fd77b056caeaf1a603ea02d711175a69bf