Create new sudo rules to facilitate monitoring airflow
The existing checks for airflow were failing because they were being run
as the nagios user. This change introduces two new sudo rules permitting
the nagios user to run the checks as the analytics user. It also changes
the nrpe command definition to use sudo.

Hosts: an-launcher1002.eqiad.wmnet,an-airflow1002.eqiad.wmnet
Bug: T307102
Change-Id: I587ec05971a3db7e311b056a5288bbfb10d9a2c9