C:helm:  make the group permissions on helm_cache configurable
An update was made recently to update the permissions on the helm_cache
directory to make them more restrictive and confine them to deployment
instead of wikidev.  however the deployment group does not exists every
where.  This Cr adds a new parameter so roles can choose which group
makes the most sense.

I have updated the roles for release and ci::master.  for ci::master
picked the group contint-users which seemed logical.  for releases i
couldn't see a logical group to choose so i have reverted this back to
wikidev

Hosts: C:helm
Bug: T305729
Bug: T307740
Change-Id: Ie1b87c9f896bcb0a51700cd0ceccf69707b61dc7