O:cache::text: move netbox-next to the cacheing infrastructure
We plan to move netbox to use dns discovery via the caching
infrastructure.  As such we also want the dev infrastructure to mirror
this set up.  This Cr adds netbox-dev2002 as the only backend for the
netbox-next.wikimedia.org domain.  i.e. no DNS discovery for the dev
instance as there is only one host

Related-To: I9794deb1350f0af1645b214348d94c963bcf2db5

Bug: T296452
Hosts: O:cache::text
Change-Id: I35e48b2f72aec2fdd52dc82c0e796602d4fda8b5