dhcpd: downgrade an-tool1005 to stretch, upgrade an-tool1007 to bullseye
The testing of turnilo on staging was successful, so it can be deployed
to an-tool1007. Superset, however, does not run on Debian Bullseye since
it lacks the python package python-gevent, and superset and turnilo are
colocated on an-tool1005. So turnilo staging won't work for the moment.
This could be resolved by making superset compatible with Bullseye or
putting turnilo staging on a different host.

Bug: T301990
Change-Id: I76130bcb6490a406112bab259fd8fa70334b53ed