durum: return the site/DC in the check response
The current durum response for a successful Wikidough DoH or DoT
connection looks like:

    {"wikidough": true, "service": "dot"}
    {"wikidough": true, "service": "doh"}

In addition to this, we should also return the site/DC, so that the user
is aware of the DC they are connecting to. It's also possible to return
the hostname in this response but that is probably not as useful as the
DC they are connecting to, so let's plan that for another commit.

After this commit, the response will look like (assuming DC is eqiad):

    {"wikidough": true, "service": "doh", "site": "eqiad"}

We will then update the check.wikimedia-dns.org frontend to display the
site in the result.

Hosts: O:durum
Change-Id: I5d011db1049be7dd920a53d9cb141b848af8fc65