durum: display the DC the user is connected to in the frontend
For the durum frontend (check.wikimedia-dns.org), display which data
center the user is connected to. This is useful information for the user
for a number of reasons: transparency, which country (or state in the
US) there queries are being forwarded to, debugging, etc.

This builds on commit d7e3c98b in which we added the JSON response that
we now format and display in the frontend.

Hosts: O:durum
Change-Id: Ib42c3b836860e9c6c70c2bdc7557a2c52828e2a7