service::docker: refresh the service when its image is updated
Puppet currently only triggers a pull of the image if it does not exist
locally. However, to support services running at `latest`, we have to
query the image registry to make sure that the local version of the
image is not stale.

Unfortunately there is no idempotent way to make this check using the
CLI. docker-pull does not have a --dry-run mode (feature request:
https://github.com/distribution/distribution/issues/2412), nor is there
another straightforward way to use the docker CLI to check if an image
is up-to-date. It's possible to query the registry via curl, but it's
not pretty : https://stackoverflow.com/a/64309017/582542 .

The solution I landed on (in which the 'unless' clause actually updates
the image, and the main exec is a dummy) is a bit ugly but it works.

Bug: T308598
Change-Id: Ifbbfef8532a0ecfac865f1b9035154113c17c2bd