deployment_server: switch the deployment group to 'deployment'
This will change ownership of various things and make only people
in the deployment group able to deploy mediawiki, as opposed to anyone
having a login on the deployment servers.

That was acceptable years ago, when the deployment servers were only used
for MediaWiki; but today it means we're granting that power unintentionally to
a bunch of other people.

Bug: T303857
Change-Id: I77cdeb7f490d66693d898e5de2bbad402aa5de3e