idp: add gitlab-replica-new to idp
gitlab-replica is a passive instance of production GitLab.
gitlab-replica-new is a new physical instance of gitlab-replica.
Login to this instance should work so integration to idp/CAS is needed.
This changes adds gitlab-replica-new to idp/CAS regex for gitlab-replica.

gitlab-replica-new is needed for migration only. After that it will be
renamed to gitlab-replica and the -new entry can be removed again.

Bug: T307142
Change-Id: Ie99e8d20fad6905ce7c9ca3cd194989466a995f2