Add new Cassandra cluster for ML cache/feature-store workloads in eqiad
The new cluster will support two things:
1) Score cache for things like ORES (implemented in Redis atm).
2) Feature store to support ML serving API.

For 1), we just need a simple keyspace in which we can store
the result of scoring a specific rev-id for a given wiki. Processing
a score with the revscoring/ORES models (that we'll run on Lift Wing,
the new infra) is very expensive (~500ms) so we need to be able to
cache/store results and Cassandra seems a good compromise between
efficiency and maintainability (expecially for traffic load balancing
etc..).

For 2), our idea is to use feast.dev, but we are still not sure.
The keyspace will likely be auto-managed by the tool, but since
the feature datasets will come from Data Engineering (so eqiad only)
we'll likely use a network topology to mirror the namespace to
the ml-cache codfw cluster (still to be created, for the moment
we'll concentrate on the eqiad one).

The first test will be a simple single-instance-per-node Cassandra
cluster, using profile::cassandra with a multi-instance hiera
config for consistency with the other clusters.

We have, for the moment, 3 nodes in eqiad and 3 in codfw, 1.4T of
disk space after raid1, 256G of ram and 48 vcores (got it from nproc).

Bug: T302232
Change-Id: I127560ea48585cc164fafdae4756e6dc4f8ac52f