gitlab: reduce backup_keep_time to 2d
GitLab virtual machines have 100gb disks for backups. With the
current backup size of around ~20gb and the unefficient way gitlab
is doing backups, this is not enough. Backups fail with no space
left on device.
This change reduces backup_keep_time from 3d to 2d temporarily, until
migration to physical hosts is finished (T307142). This will affect
backups on GitLab host, backups in bacula will be stored longer than
2 days.

Alternative solution could be to add a bigger disk for backups or
to optimize backups on our side (for example by not doing a full copy
for latest backup).

Bug: T274463
Change-Id: Icf612d8c1eb0a9770fef2fd3905521fedd38fdc3