gitlab: retry rails console, don't keep gitlab-secrets.json
This change adds a retry to the gitlab-rails console command, because
GitLab needs some time after a restart/restore. Otherwise the execution
fails with FAILED - Internal API error (502).
This change also removes the special handling of secrets file. The
replica doesn't need a dedicated secrets file, only a dedicated gitlab.rb
configuration file. If a dedicated secrets file is used, the replica is
unable to decrypt data (which are encrypted using the secrets from the
backup).

Bug: T274463
Bug: T308089
Change-Id: Ic273f650b90c098031de96c9f7b453b575bfbc70