base: remove "managed by puppet" notice on /etc/skel/.bashrc
Remove the "this file is managed by puppet" comment header on the
content provisioned at /etc/skel/.bashrc. This notice is technically
true, but also confusing for users in Toolforge and other environments
as the copy of /etc/skel/.bashrc that ends up at $HOME/.bashrc is
detached from Puppet's management of the skel directory.

Change-Id: I6d3b937b36083a20580557883345b6c3e589d07c