gitlab: rsync config and data backup to same folder on replica
Due to rsync restrictions it is not possible to sync the config
backup directly to /etc/gitlab/config-backup on the replica. This
change will sync the config backup to the same dedicated volume
(/mnt/gitlab-backup). Folders in the restore script were changed
in I251e0093e805433f854da09e126475b269206c56 already.

This change does not merge both backup locations for production.
data backup and config backup (which includes keys for decryption)
are stored in two different folders in production and have two
dedicated filesets in bacula.

Bug: T274463
Bug: T307142
Change-Id: Id05e1c64d5ef90409edeb43d26c8fb60e84bb605