wmflib::service: add data loader class
Currently each call to wmflib::service::fetch preforms a lookup of the
service:catalog.  Futher each call to get_pool_nodes calls a loadyaml
for the specific.  In theory the lookup call should be cached however im
not sure that the loadyaml call is.  As such we create a new class to
load the data on the first call of one of theses functions and then
refer to this data for each subsequent call

The spec test in Gerrit change 776225 (the next in this change) are
9x quicker.  from ~45 seconds to ~8 seconds

Change-Id: I376747b23847e601a4dd8ede45c0e3408644df8f