memcached: Untangle TLS/1.6 options
For buster a backport of memcached was introduced which also enabled TLS.
This is being used by the IDPs and the main memcached servers/gutter pool.

Now that bullseye has 1.6 by default, but no TLS by default, there's a
new TLS-enabled build, but we need to untangle the options passed to the
class for TLS and 1.6.

Hosts: C:memcached
Bug: T308214
Change-Id: I40062946689e7d5718a986535abd390361445c3e