parsoid::testing: move apache/php auto_restarts to separate profile
We have 2 types of parsoid test servers, those that include
MediaWiki and with it apache2 and php-fpm, that is scandium,
and those that are just testreduce servers and do not include
MediaWiki, that is testreduce1001.

Both have nginx installed but apache and php don't exist on the
latter.

Therefore the auto_restart units created there fail since they
have nothing to restart.

Move the "diff" between both servers into a specific profile
for the first type of servers to avoid those.

Change-Id: I7eb5160007050de0d566ddcf2281421e776b4145