gitlab: create role/profile to temp use gitlab1004 to store backups
We are currently running out of space on the production gitlab
machine and can't do a full backup without deleting all previous
backups.

This is partially because we had a sudden spike in total
size of all repos which is investigated separately.

We also have new physical machines to replace the VMs (gitlab1003
will replace gitlab1001) and we have dedicated partitions on
old and new server (this is why the service did not go down when
the last full-backup attempt filled the disk).

Additionally we are switching from /mnt to /srv and are in the
middle of adjusting backup and restore scripts.

As a temp stop gap to ensure we have backups and be able to
delete them from the prod server and start a full backup run there,
temp use gitlab1004, which so far is not assigned any role,
to store existing backups there.

Bug: T274463
Bug: T307142
Change-Id: I6f1a56bda16e9c6f5f13be1ee1d83c885dceadd2