gitlab::dump: use rsync::module directly, not quickdatacopy
Reason: It expectes the rsyncd to run on the _source_ host
to be pulled from.

We don't want this here though, we don't want to mess with gitlab1001
at all.

We just want to push to gitlab1004 from it.

What is source and what is dest is a constant source of confusion with
this.

follow-up to I6f1a56bda16e9c

Bug: T274463
Change-Id: I0eb16e28ce446f66a2a53eead91532c440ba14e2