idp: add gitlab-new to idp
gitlab is the production GitLab instance. gitlab-new is a new physical
instance of gitlab production. Login to this instance should work so
integration to idp/CAS is needed.
This changes adds gitlab-new to idp/CAS regex for gitlab.

gitlab-new is needed for migration only. After that it will be
renamed to gitlab and the -new entry can be removed again.

Similar to Ie99e8d20fad6905ce7c9ca3cd194989466a995f2

Bug: T307142
Change-Id: Ibb1fdd2221aba7712e01b1f747607a7db3442c50