Configure .gitignore to exclude the vendor subdirectory
When running `bundle install` from within the puppet repository gems are
downloaded and added to a new vendor/bundle subdirectory. This directory
is then seen as an untracked directory and clutters up working copies.

This commit adds the vendor directory to the .gitignore file, so that we
can maintain a bundle environment without this untracked directory being
shown.

Change-Id: I1ae6f25597e10b3b9789394a871320641e5b8e1f