imagecatalog: do not have auto_restart service if service does not exist
The actual server imagecatalog has the $ensure parameter passed through
and only exists on the active server. Does not exist on the passive
server, deploy2002.

But the auto_restart service for it does not pass it through and
therefore exists on all servers.

This leads to a failed auto_restart service because it can't find the
imagecatalog service.

And this leads to Icinga alerts for systemd state being broken.

Change-Id: Ie5a2a508fa18842c6bccd93927923f66525bdf96