Change cloudgw configuration to use rack-specific GW IP
There are two cloudgw hosts in Eqiad.  Cloudgw1001 is connected to
cloudsw1-c8 in rack C8, and cloudgw1002 is connected to cloudgw1-d5
in rack D5.

VRRP is currently running on Vlan1120 betweeen cloudsw1-c8 and
cloudsw1-d5, with VIP  185.15.56.241.  The cloudgw hosts both have
this IP configured as their default gateway in vrf-cloudgw,
configured in /etc/network/interfaces.d/cloudgw (see
https://w.wiki/5E5A).

I suspect the IP was previously used on the CR routers, in which case
it would make sense for the cloudgw to use the VIP.  But as the VIP
is now configured between the cloud switches, and each cloudgw is
connected to one and only one cloudsw, it does not make sense to
keep using the VRRP VIP as gateway.  If a cloudgw link, or the cloudsw
it is connected to, fails, it cannot use the other switch as GW
instead.

As there is only a 50% chance either cloudsw is VRRP master at a given
moment, using the VIP is inefficient for outbound cloudgw traffic.
Half the time cloudgw devices will send outbound traffic across rows,
rather than using the switch in their own rack as GW.

This change adjusts the configuration parameter used to build the
cloudgw interface config, so that each cloudgw uses the unique IP
of the cloudsw in its own rack as GW, rather than the VIP.

Bug: T304989
Change-Id: I4cdf4d80f27ef37a64b120b04721bfea26405f1e