Add cloudsw1-e4 and cloudsw1-f4 to mgmt and adjust existing cloudsw
The main element of this change is to add  cloudsw1-e4 and cloudsw1-f4
to management.  As they are L3 switches they are reachable both on their
loopback IPs and dedicated management port.

Also changing non-mgmt IPs for cloudsw1-c8 and cloudsw1-d5.  The old
addresses for these are now within the cloudvrf and they have new
10. loopbacks in the prod realm.

This patch also adds the cloudsw2 devices to the mgmt_switches
definitions, and changes the IP addresses defined in hierdata for
them to the ones configured on their lo0.0 interfaces.  Also switching
the type from l2sw to l3sw.

Bug: T304989
Change-Id: Ib5420956b2454d42524b007b266f7047b59d41ed