Split up the tables we sqoop
Of all the many tables we sqoop from the cloud replicas, only 6 are
currently required for the mediawiki history data pipeline.  Splitting
them up speeds up sqooping and makes it less dependent on schema changes
in the other tables.  The cost is that this is slightly harder to
maintain, so I tried to name the files something very obvious.

Bug: T309806
Change-Id: Ie0dc3dfcf9395661c77919fd70caf8224d2fbdff