wmcs-image-create: Use openstack cli for creating new glance image
The python glance client has been failing in codfw1dev when uploading
image data. I'm not sure why, but the CLI works, so this patch changes
us to use the CLI.

This is not an improvement, but it does get things working. As compensation
this patch also de-deplicates a few lines.

Change-Id: Ia840f11077e13f91ae9503c9d704f3ab2a3a522f