logstash_checker.py: Change `--delay` argument type to float
Scap has always assumed that `logstash_checker.py` would accept a float
for `--delay`, but it seems it has only recently started calculating and
and passing a float which is causing `logstash_checker.py` to fail.

Let's change the argument type of `--delay` to a float to allow it to
accept either an int (backwards compatible) or a float (what scap
expects).

Bug: T308068
Change-Id: I918879bdd119db8a0f3ff99d1b19d26562370254