C:apereo_cas: Disable u2f by default
This Cr fixes the config which is currently meaning u2f is not working.
However support for it is starting to deprecated (T296629) and with the
recent upgrade of cas we expect to start rolling out webauthn.  As such
disable u2f by default as well

Bug: T296629
Hosts: C:apereo_cas
Change-Id: I607d4120f0db44ecad87e28dde6667253f391206