scap: switch over from Debian package to self-installed scap
On scap targets, use a symlink to add the self-installed scap to the
PATH and replace the version from the deb package. The configuration
to install the deb package will be removed in a followup change.

Also piggybacked the removal of a related rsync module definition that
is not required anymore

Bug: T303559
Change-Id: Id06161a6707214a48731de776971926e8c0315f9