Decrease the retention time on the kafka-test cluster to 1 day
Previously this cluster inherited the default settings of 7 days (168
hours) for its log retention value. This has started to cause a problem
when we start working with larger topics, since the kafka-test servers
do not have a very large storage capacity.

Therefore we have decided to reduce the log retention from 7 days to 1
day on the test cluster.

Bug: T310342
Change-Id: I63733b2da8cfb2b5d1f095cb965087ed14a5c2a4