aptrepo: add repository component for bird2
For upgrading to bird2, add the component to aptrepo. The reason for
doing this is that most of the existing hosts using bird (and bird6) are
currently on buster. The bird2 package is available in buster-backports
[0] and as per the discussion with Moritz on IRC and his suggestion
(summarized below), we should add the package to main.  The reasons for
doing so:

- *-backports is only operated for the released Debian version and with
  buster moving to LTS [1] in July 2022, buster-backports may be
  deprecated at any time without a warning.

- Even though all the existing bird hosts should eventually upgrade to
  bullseye, that hasn't happened yet. bullseye has bird2, so there will
  be no issues there.

[0] - https://packages.debian.org/buster-backports/bird2
[1] - https://wiki.debian.org/LTS

Bug: T310574
Change-Id: Ie71c66d97a6f2917c6fc80ec1f902bcc42434708