OpenStack haproxy: make Galera checks half as often
This will double the time we wait before declaring a check failed, and
double the time we wait before declaring the service back up. This
will sweep some failures under the rug but will also reduce the
number of times we do a full failover between galera endpoints;
I'm hoping it'll be a net improvement.

Change-Id: Id31301d7a31ab8897f9cb04af5895d898e18ce56