Add a new check for the age of the standby namenode fsimage
This check ensures that the fsimage backup on the standby namenode is
not out of date. If it becomes stale then backups are taken of the
stale image, which makes them less useful.

Note that only the standby server writes these fsimage files. If we
perform a manual failure and run with an-master1002 as the active server
for an extended period of time, this alert will trigger as it is
currently configured.

Bug: T309649
Change-Id: If62bb2f0a9d928482a428270e57262f60c41cdc4