base: create profile to allow unprivileged userns, use it on gitlab_runners
Enable unprivileged user namespaces for rootless operation of buildkitd.

Alternative would be to run it as root but that's what we don't want on
protected runners.

Without it the buildkitd service won't run, after setting it manually it
keeps running.

https://security.stackexchange.com/questions/209529/what-does-enabling-kernel-unprivileged-userns-clone-do

As requested create a separate profile for this to make it easier to
identify servers allowing it.

Bug: T308271
Bug: T287081
Bug: T304675
Change-Id: I3efe7c3fe5f6689bd6feb290c909f03308c146fa