varnish: sort query parameters on the Beta Cluster
Sorting query parameters improves cache efficiency by preventing the
creation of duplicate cache objects for requests that are semantically
equivalent. Before we can do this for production traffic, we need to
verify that MediaWiki is insensitive to parameter order. Enabling this
on the Beta Cluster will help us verify that.

Unfortunately there is already one way in which Varnish's
std.querysort() can alter the semantics of the request, which has to do
with the way it handles duplicate parameters. Varnish retains
duplicates, but it uses the full key-value pair as the sort key, such
that '?foo=b&foo=a' gets rewritten into '?foo=a&foo=b'. There are two
ways this can break things:

1) Array parameters: PHP Decodes '?foo[]=z&foo[]=a' into ['z', 'a']. In
   this case, changing the order of parameters changes the order of
   elements in the decoded array. We handle this case by not sorting any
   URL that contains '['.

2) Non-array parameters: when a non-array parameter is present multiple
   times, the lattermost occurrence "wins": e.g. if '?foo=x&foo=y', then
   $_GET['foo'] is 'y'. This isn't part of any web standard or PHP
   spec, but it's been the practice since forever so changing it can
   alter the behavior of existing URLs, and it can break badly-written
   code. I don't think this will affect any code that has been reviewed
   but it could break some user scripts or gadgets.

So, why do this? Because Beta is low-risk and we'll get some useful
intelligence this way. When this gets rolled out to production we might
be using HAProxy which has a query sorting function with more sensible
handling of duplicate parameters.

Bug: T138093
Change-Id: I4c8e918c2279baa243e7fcd0c57fc7551a22ec74