gitlab: add prometheus blackbox http monitor
We are using the receivers added in I3d40658cf973d1096d0e.

This created the team serviceops-collab and 2 severity level
reactions.

For a WARN we should get email and IRC notifications
and for a CRIT we should get both of these and additionally
a Phabricator ticket.

The default is "critical" and does NOT mean paging in this context.

Checking for a specific string in the body of a specific URL too,
not just for a 200 on /.

Using "Explore GitLab" from the /explore page which you get
redirected to from /.

Change-Id: I6a8a12a25939a5202a1ad81945a70e31711fc4b1