prometheus: Add ipmi_exporter to bullseye+
Why:
prometheus-ipmi-export could provide us with data regarding:

* bmc information like firmware revisions
* chassis power states
* power consumption metrics
* SEL capacity and usage
* Temperature sensors
* Fan Speed sensors
* Voltage sensors
* Current sensors
* Power sensors
* Other generic sensors

What:
Create a class to install the package, make sure the service is running
and populate a sane configuration file. Unfortunately, that sane
configuration file can't be used, as we have 1.2.0 in bullseye
currently. Instead populate a set of hacks with a symlinked sudo wrapper
instead, until we upgrade. Include that profile in base.pp

Change-Id: I0c1a0b9ef2a1310fa5d0c98def92f313cb30f9e6