P:netbox: add dynamic config back to config file
Having the dynamic config in the configuration.py files means that it
is not possible to override it via the django admin portal.  It also
means that it is managed, tracked and part of our change review process
via gerrit.  As such add it back and creat hiera defaults.

In this CR we have also set the retention policy for cloud and
netbox-next to 1 as its expected theses systems will be test systems
and as such don't require to keep so much history

Hosts: C:netbox
Bug: T311048
Change-Id: I95d909e353ba03744c4938a792b6a79663a4c199