C:base::puppet move Puppet to Systemd timer
To avoid generating alerts, should the Puppet run fail,
we add "ignore_error => true". This will prefix the
ExecStart in the systemd service with "-".
From the systemd man page:
If the executable path is prefixed with "-", an exit code
of the command normally considered a failure (i.e. non-zero
exit status or abnormal exit due to signal) is recorded,
but has no further effect and is considered equivalent to
success.

Bug: T273673
Change-Id: Ia44463b833e9303389c955035dd7d2f46a578f97