openstack.vendordata: reduce timeout so it retries
Cloud-init does two different request to the metadata services, one
first quick check to see which ones are alive, and then one to
retrieve the metadata (vendordata).

For the first one, it does not use the `retries` parameter, but
instead, it will keep retrying until it reaaches `max_wait`, using the
given `timeout` for each request.
This tries to ensure that more than one request is going to be done.
For details see gh issue cloud-init#1529

Bug: T309930
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I739833e85e88de3432ff47ce0ee8a84eb49e694d