Keyholder: add new agent for trainbranchbot
When running the MediaWiki train, we run scap deploy-promote which
updates MediaWiki versions in wikiversions.json.

The push is currently done with individual users' personal account.

We would like to do this with a shared user, automatically.

This user needs Gerrit credentials, which is what this patch adds.

Note: This will require root to add a new ssh keypair to puppet secrets
before this patch merges.

Bug: T310620
Change-Id: Ic148b76ecc35bc8c3917878d09bdd73b4020f80f