query_service: fix syntax error in apache config
In Icab18579bd4ffecc21e4059d88be522a9550f00b we added this new
FilesMatch directive, but it was accidentally closed as Files instead of
FilesMatch.

Just for context, note that this httpd.erb file is rendered to
/etc/apache2/sites-enabled/50-commons-query-wikimedia-org.conf
whereas actual requests for the /sparql endpoint are routed to the nginx
that exists on the wdqs host itself.

Bug: T289243
Bug: T301461
Change-Id: Iae72f0a2b94ff96edd38174d13a22fb367e2902a