alertmanager: create receivers for serviceops-collab
Create 2 new 'receivers' to be used with blackbox monitoring
like I6a8a12a25939a5202a1ad.

for services: gitlab, gerrit, phabricator, vrts etc handled
by "serviceops-collab" subteam.

One receiver is for severity=critical, this is supposed to
send IRC notifications, email notifications AND create a ticket.

A second one is for severity=warning, this is supposed to send
IRC and email notifications but not open tickets.

As opposed to other teams we name the receiver after the severity
not after the notification method since we want multiple
methods per severity level.

We should discuss if this is really the way we prefer.

Delibarately not the same as "core serviceops" but so far
using the general "serviceops" tag in Phabricator. We might
want to create our own there too.

Change-Id: I3d40658cf973d1096d0ece01bdf206400d5209fd