Raise profile::cumin::monitoring_agentrun::crit
With almost 2000 servers there's always a certain amount of hosts which
currently in setup/broken or other situations which makes this trigger.
As a consequence this alert is permanently in critical state, which doesn't
seem useful. Bump it to 20, we still want to see an alert if there's some
larger fleet-wide issue.

Change-Id: I0107e73f02df60ece4554ee63a1614f821ad33ba