Increase the java heap for the Hadoop namenodes again
We have continued to see failures of the Hadoop namenodes to failover
and fail back successfully, due to garbage collection pauses triggering
the health check mechanism.

The hosts still have RAM that is unused, so we can increase the heap
size again. This commit increases the heap size from 72 GB to 84 GB.

Bug: T310293
Change-Id: I4d6ac7838db7040750e02cd490c7c8339d946217