Add a new partman recipe for the new H750 based stat servers
This adds a new partman recipe that uses hardware RAID on the new H750
based servers. The O/S is installed to a pair of SFF disks in the flex
bays, designated as /dev/sdb. The other device is a RAID 10 volume using
the primary LFF bays and designated /dev/sda.

Both devices use LVM, are formatted, and allocate 95% of the space. The
primary storage is formatted as ext4 and mounted to /srv/

This configuration is applied to two new stats servers, stat1009 and
stat1010.

Bug: T299466
Bug: T307399
Change-Id: Ib18b80ec2dd4e9b6dd325a67eebeeac7a3a6a78a