ipmi_exporter: Have wrapper depend on sudo::user
Why:
There exists a race condition when (re)-imaging a host where the service
might start (after the package gets installed), the wrapper is
populated, but sudo policies aren't there yet. A couple of minutes
later, as the puppet run ends, everything will be ok, but the time
window is sufficient for unauthorized sudo alerts to be sent out.

What:
Have both the config file and the wrapper depend on the sudo::user
define

Change-Id: Idaeb54ea7bcc60fb8bfe25f0ed973fa39979c6f2