admin: add gitlab-roots group to gitlab_runner role
Recently a change was made to include contint-roots members
in the gitlab-roots group to give them access to gitlab servers.

The gitlab-roots group has so far been applied on gitlab servers
but not on gitlab runners. So only root users have access so far.

Therefore contint-roots do not have access to gitlab-runners yet
but the request and assumption was that they could.

Bug: T308350
Change-Id: I8f6054953dc9a270843dc384af33277d5c0cd5e0