Update the partman recipe for use with the new stat servers
We have been using stat1010 as an experimental host for testing the new
PERC H750 controller, which was exhibiting some unexpected behaviour by
swapping device names of /dev/sda and /dev/sdb

Now that the behaviour is understood and can be reversed, we can use a
more generic partman recipe for this type of server. For now we will
apply an existing recipe for kafka-jumbo to these servers to continue
testing. If this testing is successful we will harmonize the recipe
names at a later date.

Bug: T307399
Bug: T297913
Change-Id: I873e474dbfca190120967db45ea48ac086507372