Reduce the minimum size of /srv in the kafka-jumbo recipe
This recipe was previously created for a 12 x 4 TB RAID10 /srv volume,
but we are now working towards making this a more flexible recipe,
suitable for any host that uses hardware RAID and only has a large /srv
volume on /dev/sdb

The host we are trying it on at the moment (stat1010) has only 4 x 4TB
drives in a RAID10 configuration. Therefore with 80% of the volume group
used, a 6TB minimum value should fit.

Bug: T307399
Change-Id: I47e390c34345a081b94eb12daeb943937caa4fcf