striker: connect docker container directly to host network
Allow Striker running inside a Docker container to communicate with the
nutcracker process attached to 127.0.0.1 of the host computer by
connecting Striker's container directly to the host computer's network.
Use of `--network host` with `docker run` provides the same level of
network isolation as if Striker were running directly on the host and
not in a container. All other Docker managed isolation--storage, process
namespace, user namespace--are preserved.

* Add a $host_network parameter to service::docker which allows
  configuring the container to attach to the host network rather than
  the default bridge network.
* Set $host_network => true for Striker's service::docker instance.

Bug: T306469
Change-Id: I9959289901b25f469d5cb1b2b1ccbd6d06bae8e1