wmcs-enc-cli.py: fix args passed to requests.post
The requests session is /almost/ a drop-in replacement for
the direct call on requests() but the session wants
named args rather than positional.

This seems to not mess up the formatting of the hiera
that gets stored but I wouldn't bet my life on it.

Follow-up to https://gerrit.wikimedia.org/r/c/operations/puppet/+/779899

Bug: T274666
Change-Id: I588497f4ccccfd5ea1d4df7dbe5be22dae5a253f