mediawiki::php: fix proxy selection when using unix sockets
Given the fact that httpd's mod_proxy_fcgi only matches what is
after the pipe to identify a worker, using the same fcgi://localhost slug
for all proxies wouldn't work.

So, change the name of all the additional workers to be something like
"fcgi://www-7.4" for all non-standard versions.

Also fix a naming issue with the fcgi socket in the php class.

Change-Id: I4067b7c970293ce552808a147887a2bc24aefacf