vrts: add promtheus blackbox monitoring
This adds a monitoring check for VRTS (formerly OTRS)
for the page:

https://ticket.wikimedia.org/otrs

content "wikimedia".

It uses the recently added serviceopss-collab receivers so this
will mean if VRTS goes down we should be alerted via IRC, email
and an automatically created Phabricator ticket.

I just copied this from what I did and got reviewed for gitlab.
Found the linked ticket afterwards.

Bug: T303190
Change-Id: Ifc8a83f46ef726eb501ca3e0fc94eefb5ffe9056